Implementación del CRUD, Layout, etc

+++ database/migrations/2022_03_16_161901_create_personas_table.php
@@ -17,16 +17,16 @@ class CreatePersonasTable extends Migration
             $table->id();
             $table->string('nombre');
             $table->integer('edad');
-            $table->float('peso', 8, 2);
-            $table->float('estatura', 8, 2);
-            $table->float('masa_muscular', 8, 2);
-            $table->float('grasa_visceral', 8, 2);
-            $table->float('grasa_corporal', 8, 2);
+            $table->double('peso', 8, 2);
+            $table->double('estatura', 8, 2);
+            $table->double('masa_muscular', 8, 2);
+            $table->double('grasa_visceral', 8, 2);
+            $table->double('grasa_corporal', 8, 2);
             $table->integer('edad_metabolica');
             $table->integer('diferencia_de_la_edad');
-            $table->float('medida_de_brazo', 8, 2);
-            $table->float('medida_de_pecho', 8, 2);
-            $table->float('medida_de_pierna', 8, 2);
+            $table->double('medida_de_brazo', 8, 2);
+            $table->double('medida_de_pecho', 8, 2);
+            $table->double('medida_de_pierna', 8, 2);
             $table->timestamps();
         });
     }
